# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Yarn Dyeing Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the returnTo path on success.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
...
  /factories:
    get:
      tags: [factories]
      operationId: listFactories
      summary: List all factories
      parameters:
        - name: includeArchived
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: List of factories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Factory'
    post:
      tags: [factories]
      operationId: createFactory
      summary: Create a new factory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FactoryInput'
      responses:
        '201':
          description: Factory created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Factory'
  /factories/{id}:
    get:
      tags: [factories]
      operationId: getFactory
      summary: Get a factory by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Factory found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Factory'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [factories]
      operationId: updateFactory
      summary: Update a factory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FactoryUpdate'
      responses:
        '200':
          description: Factory updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Factory'
    delete:
      tags: [factories]
      operationId: archiveFactory
      summary: Archive a factory (soft delete)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Factory archived
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Factory'

  # ── ORDERS ─────────────────────────────────────────────────────────────────
...
  /dashboard/recent-activity:
    get:
      tags: [dashboard]
      operationId: getRecentActivity
      summary: Get recent order status changes
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Recent activity list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActivityItem'

  # ── REPORTS ────────────────────────────────────────────────────────────────
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'

    # Storage schemas
...
    Factory:
      type: object
      required: [id, factoryCode, name, status, createdAt]
      properties:
        id:
          type: integer
        factoryCode:
          type: string
        name:
          type: string
        address:
          type: [string, 'null']
        contactPerson:
          type: [string, 'null']
        phone:
          type: [string, 'null']
        email:
          type: [string, 'null']
        status:
          type: boolean
        createdAt:
          type: string
          format: date-time
    FactoryInput:
      type: object
      required: [factoryCode, name]
      properties:
        factoryCode:
          type: string
        name:
          type: string
        address:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
    FactoryUpdate:
      type: object
      properties:
        factoryCode:
          type: string
        name:
          type: string
        address:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        status:
          type: boolean

    # Order schemas
...
    ActivityItem:
      type: object
      required: [id, orderNo, buyerName, status, createdAt]
      properties:
        id:
          type: integer
        orderNo:
          type: string
        buyerName:
          type: string
        factoryName:
          type: [string, 'null']
        status:
          type: string
        createdAt:
          type: string
          format: date-time

    # Report schemas
...