# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Yarn Dyeing Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: number;
  orderNo: string;
  buyerName: string;
  /** @nullable */
  factoryName?: string | null;
  status: string;
  createdAt: Date;
}