# lib/api-zod/src/generated/types/factory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Yarn Dyeing Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Factory {
  id: number;
  factoryCode: string;
  name: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  contactPerson?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  status: boolean;
  createdAt: Date;
}