# lib/api-zod/src/generated/types/factoryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Yarn Dyeing Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface FactoryUpdate {
  factoryCode?: string;
  name?: string;
  address?: string;
  contactPerson?: string;
  phone?: string;
  email?: string;
  status?: boolean;
}